🚚 Rename: 폴더명 소문자로 변경 (#166)
