+++ src/components/my-page/bookmarked-pages/BookmarkedLecture.tsx
@@ -1,6 +1,6 @@
 import { ListItemSkeleton } from '@/components'
 import { Input } from '@/components/common/input'
-import EmptyDataState from '@/components/common/stateTemp/EmptyDataState'
+import EmptyDataState from '@/components/common/state/EmptyDataState'
 import BookmarkedLectureCard from '@/components/my-page/bookmarked-lecture/BookmarkedLectureCard'
 import type { BookmarkedLectures } from '@/types/api-response-types/lecture-response-type'
 import type { UseQueryResult } from '@tanstack/react-query'
